+++ node/Cargo.toml
@@ -7,73 +7,73 @@ license.workspace = true
 homepage.workspace = true
 repository.workspace = true
 edition.workspace = true
-
-build = "build.rs"
 publish = false
 
-[dependencies]
-clap = { workspace = true }
-log = { workspace = true }
-codec = { workspace = true }
-serde = { workspace = true }
-jsonrpsee = { workspace = true }
-futures = { workspace = true }
-serde_json = { workspace = true }
-
-# Local
-parachain-template-runtime = { workspace = true }
-
-# Substrate
-frame-benchmarking = { workspace = true }
-frame-benchmarking-cli = { workspace = true }
-pallet-transaction-payment-rpc = { workspace = true }
-sc-basic-authorship = { workspace = true }
-sc-chain-spec = { workspace = true }
-sc-cli = { workspace = true }
-sc-client-api = { workspace = true }
-sc-offchain = { workspace = true }
-sc-consensus = { workspace = true }
-sc-executor = { workspace = true }
-sc-network = { workspace = true }
-sc-network-sync = { workspace = true }
-sc-rpc = { workspace = true }
-sc-service = { workspace = true }
-sc-sysinfo = { workspace = true }
-sc-telemetry = { workspace = true }
-sc-tracing = { workspace = true }
-sc-transaction-pool = { workspace = true }
-sc-transaction-pool-api = { workspace = true }
-sp-api = { workspace = true }
-sp-block-builder = { workspace = true }
-sp-blockchain = { workspace = true }
-sp-consensus-aura = { workspace = true }
-sp-core = { workspace = true }
-sp-keystore = { workspace = true }
-sp-io = { workspace = true }
-sp-runtime = { workspace = true }
-sp-timestamp = { workspace = true }
-substrate-frame-rpc-system = { workspace = true }
-substrate-prometheus-endpoint = { workspace = true }
-
-# Polkadot
-polkadot-cli = { workspace = true }
-polkadot-primitives = { workspace = true }
-xcm = { workspace = true }
-
-# Cumulus
-cumulus-client-cli = { workspace = true }
-cumulus-client-collator = { workspace = true }
-cumulus-client-consensus-aura = { workspace = true }
-cumulus-client-consensus-common = { workspace = true }
-cumulus-client-consensus-proposer = { workspace = true }
-cumulus-client-service = { workspace = true }
-cumulus-primitives-core = { workspace = true }
-cumulus-primitives-parachain-inherent = { workspace = true }
-cumulus-relay-chain-interface = { workspace = true }
-color-print = { workspace = true }
+build = "build.rs"
 
 [build-dependencies]
-substrate-build-script-utils = { workspace = true }
+substrate-build-script-utils.workspace = true
+
+[dependencies]
+clap.workspace = true
+codec.workspace = true
+color-print.workspace = true
+futures.workspace = true
+jsonrpsee.workspace = true
+log.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+
+# Local
+parachain-template-runtime.workspace = true
+
+# Substrate
+frame-benchmarking.workspace = true
+frame-benchmarking-cli.workspace = true
+pallet-transaction-payment-rpc.workspace = true
+sc-basic-authorship.workspace = true
+sc-chain-spec.workspace = true
+sc-cli.workspace = true
+sc-client-api.workspace = true
+sc-consensus.workspace = true
+sc-executor.workspace = true
+sc-network.workspace = true
+sc-network-sync.workspace = true
+sc-offchain.workspace = true
+sc-rpc.workspace = true
+sc-service.workspace = true
+sc-sysinfo.workspace = true
+sc-telemetry.workspace = true
+sc-tracing.workspace = true
+sc-transaction-pool.workspace = true
+sc-transaction-pool-api.workspace = true
+sp-api.workspace = true
+sp-block-builder.workspace = true
+sp-blockchain.workspace = true
+sp-consensus-aura.workspace = true
+sp-core.workspace = true
+sp-keystore.workspace = true
+sp-io.workspace = true
+sp-runtime.workspace = true
+sp-timestamp.workspace = true
+substrate-frame-rpc-system.workspace = true
+substrate-prometheus-endpoint.workspace = true
+
+# Polkadot
+polkadot-cli.workspace = true
+polkadot-primitives.workspace = true
+xcm.workspace = true
+
+# Cumulus
+cumulus-client-cli.workspace = true
+cumulus-client-collator.workspace = true
+cumulus-client-consensus-aura.workspace = true
+cumulus-client-consensus-common.workspace = true
+cumulus-client-consensus-proposer.workspace = true
+cumulus-client-service.workspace = true
+cumulus-primitives-core.workspace = true
+cumulus-primitives-parachain-inherent.workspace = true
+cumulus-relay-chain-interface.workspace = true
 
 [features]
 default = []
